fix: Fix a bug with absolute paths
